Revert "restructure project"
This reverts commit a893bc76aa6d10ea29c0fdf0d7bf5945f1e63cdb.
