+++ ai-training/models/prototype-V1_terminal_19_05_2023 11_23_45.h5
